+++ amd/src/booking.js
@@ -18,7 +18,7 @@
  * for all session booking and logentry events.
  *
  * @module     local_booking/booking
- * @author     Mustafa Hajjar (mustafahajjar@gmail.com)
+ * @author     Mustafa Hajjar (mustafa.hajjar)
  * @copyright  BAVirtual.co.uk © 2024
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */