+++ classes/local/subscriber/data_access/subscriber_vault_interface.php
@@ -30,7 +30,7 @@ defined('MOODLE_INTERNAL') || die();
  * Interface for the subscriber_vault class
  *
  * @package    local_booking
- * @author     Mustafa Hajjar (mustafahajjar@gmail.com)
+ * @author     Mustafa Hajjar (mustafa.hajjar)
  * @copyright  BAVirtual.co.uk © 2021
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */