+++ classes/local/views/manage_action_bar.php
@@ -25,7 +25,7 @@ use moodle_page;
  * Class manage_action_bar - Display the action bar
  *
  * @package    local_booking
- * @author     Mustafa Hajjar (mustafahajjar@gmail.com)
+ * @author     Mustafa Hajjar (mustafa.hajjar)
  * @copyright  BAVirtual.co.uk © 2023
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
@@ -220,7 +220,7 @@ class manage_action_bar extends base_action_bar {
  * Class for outputing renderable simple text label in the tertiary navigation bar.
  *
  * @package    local_booking
- * @author     Mustafa Hajjar (mustafahajjar@gmail.com)
+ * @author     Mustafa Hajjar (mustafa.hajjar)
  * @copyright  BAVirtual.co.uk © 2023
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */