+++ classes/output/views/calendar_view.php
@@ -22,7 +22,7 @@ use local_booking\external\availability_week_exporter;
  * Class to output calendar view.
  *
  * @package    local_booking
- * @author     Mustafa Hajjar (mustafahajjar@gmail.com)
+ * @author     Mustafa Hajjar (mustafa.hajjar)
  * @copyright  BAVirtual.co.uk © 2023
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */